feat(multi object tracker): add vehicle update strategy for improved tracking accuracy
