+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/tracker/update/vehicle_update_strategy.hpp
@@ -1,0 +1,45 @@
+// Copyright 2020 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__UPDATE__VEHICLE_UPDATE_STRATEGY_HPP_
+#define AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__UPDATE__VEHICLE_UPDATE_STRATEGY_HPP_
+
+#include "autoware/multi_object_tracker/types.hpp"
+
+namespace autoware::multi_object_tracker
+{
+
+enum class UpdateStrategyType { FRONT_WHEEL_UPDATE, REAR_WHEEL_UPDATE, WEAK_UPDATE };
+
+struct UpdateStrategy
+{
+  UpdateStrategyType type;
+  geometry_msgs::msg::Point anchor_point;  // used for FRONT_WHEEL_UPDATE and REAR_WHEEL_UPDATE
+};
+
+// Determines whether to use front-wheel, rear-wheel, or weak update
+// by finding the closest edge pair between measurement and prediction.
+UpdateStrategy determineUpdateStrategy(
+  const types::DynamicObject & measurement, const types::DynamicObject & prediction);
+
+// Blends measurement position/orientation into pred using a distance-weighted scheme.
+// When enlarge_covariance=true, inflates pose/velocity covariances for the weak-update path.
+void createPseudoMeasurement(
+  const types::DynamicObject & meas, types::DynamicObject & pred,
+  const autoware_perception_msgs::msg::Shape & tracker_shape,
+  const bool enlarge_covariance = false);
+
+}  // namespace autoware::multi_object_tracker
+
+#endif  // AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__UPDATE__VEHICLE_UPDATE_STRATEGY_HPP_
